+++ ak_library_management/__manifest__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 {
     "name": "Library",
-    "version": "18.0.1.9.0",
+    "version": "18.0.1.10.0",
     "license": "Other proprietary",
     "depends": ["sale_management","hr","stock","base_automation"],
     "author":"Raj",
@@ -20,7 +20,7 @@
         "views/bulk_upload_book.xml",
         "views/res_users_views.xml",
         "views/res_partner_views.xml",
-        # "views/res_config_settings_viwes.xml",
+        "views/res_config_settings_viwes.xml",
         "views/sale_order_views.xml",
         "views/borrow_transaction_history_views.xml",
         "views/stock_warehouse_views.xml",